update utils to use tape
