+++ lib/utils.js
@@ -67,7 +67,7 @@ function Utils() {
     utils.toDatabaseRecord = function(feature, dataset) {
         if (feature.id === 0) feature.id = '0';
         var f = feature.id ? _.clone(feature) : _.extend({}, feature, { id: cuid() });
-        var primary = f.id;
+        f.id = ''+f.id;
 
         if (!f.geometry) {
             throw new Error('Unlocated features can not be stored.');
@@ -82,7 +82,7 @@ function Utils() {
         }
 
         var buf = utils.encodeFeature(f);
-        var databaseFeature = utils.createFeatureKey(dataset, primary);
+        var databaseFeature = utils.createFeatureKey(dataset, f.id);
         databaseFeature.val = buf;
         databaseFeature.size = buf.length;
